Reorganizes file structure to include pages and components. Updates file paths for all changes.
